Update Android target to use new build system integration
Currently the Android build support relies on standalone toolchains which
are obsolete as of NDK r19. This change adds support for the new build
system integration described here:

https://developer.android.com/ndk/guides/other_build_systems

This avoids developers needing to create a standalone toolchain for each
target as well as needing to specify the NDK path on each target, as the
NDK path is now specified via [build]. The target-specific android-ndk
option is still available for users of standalone toolchains.