[rustdoc-json] Partially remove `paths` and introduce `external_index`
Introduce a new `external_index` that only contains a `ExternalItem`
struct from the previous `ItemSummary`. This "new" index only contains
external item and is only here so that IDs continue to referring to
an "item".

A previous version of this patch also removed `path` but at least until
https://github.com/rust-lang/rust/issues/99513 is fixed it is
unfortunately still required (as a workaround).