New macro: `pdbg!` (pretty debug)
The main intended use-case of this new macro is to relieve the user from having to choose between writing long boilerplate (`println!`) and having a messy output (`dbg!`). 

This new macro may be helpful especially when exploring new Rust features (for example in Rust playground) for educational purposes to understand how the feature works. It's modeled after the existing `dbg!` macro, except that it skips the (in some situations) redundant and distracting info about file name and line number. Another difference is that the new `pdbg!`macro will insert a new line at the end to make it easier for the user to identify where each output begins and ends at a quick glance, which may be helpful for example when outputting an enum or a struct, or some other multi-line output.