Remove `ptr::eq()` from `Cell::swap()`
It looks redundant to me, since we use `ptr::swap()` and not `swap_nonoverlapping()` so even if they are the same cell it's fine, no?

I think we could also leave the comparison and use `swap_nonoverlapping()` but I'm less sure about that, since maybe we have a valid way to obtain partially overlapping cells? I can't think of such way, but I am not sure.