Enable internal feature when testing Clippy
Clippy has internal lints, that only get build and tested, when the
`internal` feature is enabled. Those tests are just regular UI tests,
but split out from the normal UI tests, because they usually don't have
to be touched when changing something in Clippy. However, changes to
Rust internals often require updates of the internal lints, just as for
the Rest of the Clippy codebase. Currently I fix this fallout during the
sync. But this can lead to sync blockers in rare occasions.