Improve performance of `slice::sort_unstable`
At it's core it leverages branchless compare and swap operations employed in
optimal sorting networks. A variety of strategies is used to optimize for hot
and cold runtime, binary size, and max comparisons done. Many patterns see a
reduction in average comparisons performed. So this improvement is applied to
all types that are deemed cheap to move. It copies parts of the stable sort.

Before this get's merged they ought to be unified.