Implement unification of const abstract impls
If two impls cover the entire set of possibilities:

i.e.
```
impl TraitName for Struct<true> {}
impl TraitName for Struct<false> {}
```
Then it would be expected that:
```
... <const N: usize> ... where Struct<{N > 1}>: TraitName {
```
Should compile.
This allows for such by generating a pseudo-impl for
conditions which are exhaustive.

Since these impls will never be instantiated as is, it is fine to have a fake impl.

For now, this checks for a single bool condition, and whether both bool implementations are
present.

Change some of the api surface

This adds separate functionality for checking for recursive traits versus exhaustive traits.
But is still a WIP.